results+bugfixes: v3 MQ-GoRA dual-dataset run + path alignment
BUGFIXES:
  manifold_teacher.py: y.astype(int) for one-hot in classification teacher
  meta_learner.py: LabelContextEncoder auto-one-hot for 3D int lbl_nei (n_out>1)

PATH ALIGNMENT:
  Merged claude/funny-davinci → feature/gora-joint-knn (fast-forward)
  All v3 src + scripts now at /experiments/gora_tabular/ (main repo path)

CALIFORNIA (RMSE): B1=0.4433  G2=0.4546  G7=0.4929  G8=0.5169  G9=0.5099  G10=0.5208
  All v3 models worse than G2; G8-G10 collapse at epoch ~30 (label ctx overfitting)

MNIST (Acc): B1=0.9580  G2=0.9300  G7=0.9267  G8=0.9307  G9=0.9347  G10=0.9380
  G10 best v3 (+0.8% vs G2); G8→G9→G10 monotonic improvement (ablation gates ✅)
  G10 H1 specialises to FULL view (head diversity partially restored)

KEY FINDING: LabelContextEncoder works for classification (one-hot, bounded) but
  overfits on regression (unnormalized continuous targets). Teacher z_i converges
  on MNIST (loss=0.12) but not CA (loss=0.65). v4 fix: normalise y + L_agree-only.

+++ experiments/gora_tabular/src/manifold_teacher.py
@@ -91,7 +91,7 @@ def _precompute_label_centroids(
                 y_bar[i, vi] = (wts * y_vals).sum() / w_sum
     else:
         y_bar = np.zeros((N, M, n_classes), dtype=np.float32)
-        y_onehot = np.eye(n_classes, dtype=np.float32)[y]   # [N, C]
+        y_onehot = np.eye(n_classes, dtype=np.float32)[y.astype(int)]   # [N, C]
         for i in range(N):
             for vi in range(M):
                 wts = edge_wts[i, :, vi] * view_mask[i, :, vi]   # [P]
